+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- Add `switch` action to `aptly_publish` resource
+
 ## 3.0.0 - *2021-09-02*
 
 - Set unified_mode to true for Chef 17 support